Refactor test structure and add support modules for dashboard API
This commit reorganizes the test files for the dashboard API by introducing new support modules, including `support_api_harness`, `support_dashboard`, `support_graphql_knowledge`, `support_graphql_monorepo`, and `support_storage`. These modules provide utility functions for seeding test repositories and managing test data, enhancing the maintainability and clarity of the test suite. Additionally, several unused functions have been removed to streamline the codebase. The changes aim to improve the overall structure of the tests and facilitate easier updates in the future.
